+++ src/js/core/Config.js
@@ -6,7 +6,7 @@ const Config = {
 
     appName: "TrailBook",
 
-    version: "0.2.0",
+    version: "0.3.0",
 
     debug: true