+++ src/js/models/TrackSegment.js
@@ -1,0 +1,11 @@
+/**
+ * A GPX track segment kept separate from other segments.
+ */
+export default class TrackSegment {
+
+    constructor() {
+
+        this.points = [];
+    }
+
+}